Find external resources in index
